修正predict_bar

+++ candlelite/calculate/interval.py
@@ -13,7 +13,7 @@ def get_interval(bar: str, MINUTE_BAR_INTERVAL=60000) -> float:
     bar_int = int(bar[0:-1].strip())
     suffix = bar[-1].lower()
     if suffix == 's':
-        interval = (MINUTE_BAR_INTERVAL / 60) * bar_int
+        interval = int((MINUTE_BAR_INTERVAL / 60) * bar_int)
     elif suffix == 'm':
         interval = MINUTE_BAR_INTERVAL * bar_int
     elif suffix == 'h':
